fix: test build fail
