+++ bucketeer/src/main/kotlin/io/bucketeer/sdk/android/internal/di/InteractorModule.kt
@@ -1,11 +1,10 @@
 package io.bucketeer.sdk.android.internal.di
 
-import android.content.SharedPreferences
 import android.os.Handler
 import io.bucketeer.sdk.android.internal.Clock
 import io.bucketeer.sdk.android.internal.IdGenerator
 import io.bucketeer.sdk.android.internal.evaluation.EvaluationInteractor
-import io.bucketeer.sdk.android.internal.evaluation.db.EvaluationDao
+import io.bucketeer.sdk.android.internal.evaluation.storage.EvaluationStorage
 import io.bucketeer.sdk.android.internal.event.EventInteractor
 import io.bucketeer.sdk.android.internal.event.db.EventDao
 import io.bucketeer.sdk.android.internal.remote.ApiClient
@@ -15,15 +14,13 @@ internal class InteractorModule(
 ) {
   fun evaluationInteractor(
     apiClient: ApiClient,
-    evaluationDao: EvaluationDao,
-    sharedPreferences: SharedPreferences,
+    evaluationStorage: EvaluationStorage,
     idGenerator: IdGenerator,
     featureTag: String,
   ): EvaluationInteractor {
     return EvaluationInteractor(
       apiClient = apiClient,
-      evaluationDao = evaluationDao,
-      sharedPrefs = sharedPreferences,
+      evaluationStorage = evaluationStorage,
       idGenerator = idGenerator,
       mainHandler = mainHandler,
       featureTag = featureTag,
