+++ bucketeer/src/main/kotlin/io/bucketeer/sdk/android/internal/evaluation/EvaluationInteractorBK.kt
@@ -1,11 +1,13 @@
 package io.bucketeer.sdk.android.internal.evaluation
 
+import android.annotation.SuppressLint
+import android.content.SharedPreferences
 import android.os.Handler
 import androidx.annotation.VisibleForTesting
 import io.bucketeer.sdk.android.BKTClient
-import io.bucketeer.sdk.android.BKTException
+import io.bucketeer.sdk.android.internal.Constants
 import io.bucketeer.sdk.android.internal.IdGenerator
-import io.bucketeer.sdk.android.internal.evaluation.storage.EvaluationStorage
+import io.bucketeer.sdk.android.internal.evaluation.db.EvaluationDao
 import io.bucketeer.sdk.android.internal.logd
 import io.bucketeer.sdk.android.internal.loge
 import io.bucketeer.sdk.android.internal.model.Evaluation
@@ -19,9 +21,10 @@ import io.bucketeer.sdk.android.internal.remote.UserEvaluationCondition
  *
  * All methods in this class must be sync to keep things simple.
  */
-internal class EvaluationInteractor2(
+internal class EvaluationInteractorBK(
   private val apiClient: ApiClient,
-  private val evaluationStorage: EvaluationStorage,
+  private val evaluationDao: EvaluationDao,
+  private val sharedPrefs: SharedPreferences,
   private val idGenerator: IdGenerator,
   featureTag: String,
   private val mainHandler: Handler,
@@ -35,18 +38,24 @@ internal class EvaluationInteractor2(
 
   @VisibleForTesting
   internal var currentEvaluationsId: String
-    get() = evaluationStorage.currentEvaluationsId
+    get() = sharedPrefs.getString(Constants.PREFERENCE_KEY_USER_EVALUATION_ID, "") ?: ""
 
+    @SuppressLint("ApplySharedPref")
     set(value) {
-      evaluationStorage.currentEvaluationsId = value
+      sharedPrefs.edit()
+        .putString(Constants.PREFERENCE_KEY_USER_EVALUATION_ID, value)
+        .commit()
     }
 
   @VisibleForTesting
   internal var featureTag: String
-    get() = evaluationStorage.featureTag
+    get() = sharedPrefs.getString(Constants.PREFERENCE_KEY_FEATURE_TAG, "") ?: ""
 
-    set(value) {
-      evaluationStorage.featureTag = value
+    @SuppressLint("ApplySharedPref")
+    private set(value) {
+      sharedPrefs.edit()
+        .putString(Constants.PREFERENCE_KEY_FEATURE_TAG, value)
+        .commit()
     }
 
   // https://github.com/bucketeer-io/android-client-sdk/issues/69
@@ -54,15 +63,25 @@ internal class EvaluationInteractor2(
   // The server will return in the get_evaluations response (UserEvaluations.CreatedAt),
   // and it must be saved in the client
   @VisibleForTesting
-  internal val evaluatedAt: String
-    get() = evaluationStorage.evaluatedAt
+  internal var evaluatedAt: String
+    get() = sharedPrefs.getString(Constants.PREFERENCE_KEY_EVALUATED_AT, "0") ?: "0"
+
+    @SuppressLint("ApplySharedPref")
+    set(value) {
+      sharedPrefs.edit()
+        .putString(Constants.PREFERENCE_KEY_EVALUATED_AT, value)
+        .commit()
+    }
 
   @VisibleForTesting
   internal var userAttributesUpdated: Boolean
-    get() = evaluationStorage.userAttributesUpdated
+    get() = sharedPrefs.getBoolean(Constants.PREFERENCE_KEY_USER_ATTRIBUTES_UPDATED, false)
 
-    set(value) {
-      evaluationStorage.userAttributesUpdated = value
+    @SuppressLint("ApplySharedPref")
+    private set(value) {
+      sharedPrefs.edit()
+        .putBoolean(Constants.PREFERENCE_KEY_USER_ATTRIBUTES_UPDATED, value)
+        .commit()
     }
 
   init {
@@ -123,7 +142,7 @@ internal class EvaluationInteractor2(
           val updatedEvaluations = response.evaluations.evaluations
           // We will use `featureId` to filter the data
           // Details -> https://github.com/bucketeer-io/android-client-sdk/pull/88/files#r1333847962
-          val currentEvaluationsByFeaturedId = evaluationStorage.get().associateBy { it.featureId }.toMutableMap()
+          val currentEvaluationsByFeaturedId = evaluationDao.get(user.id).associateBy { it.featureId }.toMutableMap()
           // 1- Check the evaluation list in the response and upsert them in the DB if the list is not empty
           updatedEvaluations.forEach { evaluation ->
             currentEvaluationsByFeaturedId[evaluation.featureId] = evaluation
@@ -135,21 +154,20 @@ internal class EvaluationInteractor2(
           shouldNotifyListener = updatedEvaluations.isNotEmpty() || archivedFeatureIds.isNotEmpty()
         }
 
-        // 3- Save `currentEvaluations` to the database &
-        // Save the UserEvaluations.CreatedAt in the response as evaluatedAt in the SharedPreferences
-        try {
-          evaluationStorage.deleteAllAndInsert(currentEvaluations, evaluatedAt)
-        } catch (ex: Exception) {
+        // save `currentEvaluations` to the database
+        val success = evaluationDao.deleteAllAndInsert(user.id, currentEvaluations)
+        if (!success) {
           loge { "Failed to update latest evaluations" }
-          GetEvaluationsResult.Failure(
-            BKTException.IllegalStateException("error: ${ex.message}"),
-            featureTag,
-          )
           return result
         }
+        // directly update the in-memory cache
+        evaluations[user.id] = currentEvaluations
 
         this.currentEvaluationsId = newEvaluationsId
         userAttributesUpdated = false
+
+        // 3- Save the UserEvaluations.CreatedAt in the response as evaluatedAt in the SharedPreferences
+        evaluatedAt = response.evaluations.createdAt
 
         // Update listeners should be called on the main thread
         // to avoid unintentional lock on Interactor's execution thread.
@@ -168,7 +186,7 @@ internal class EvaluationInteractor2(
   }
 
   fun refreshCache(userId: String) {
-
+    evaluations[userId] = evaluationDao.get(userId)
   }
 
   fun clearCurrentEvaluationsId() {
@@ -190,8 +208,7 @@ internal class EvaluationInteractor2(
   }
 
   fun getLatest(userId: String, featureId: String): Evaluation? {
-    return evaluationStorage.getBy(featureId)
+    val evaluations = evaluations[userId] ?: return null
+    return evaluations.firstOrNull { it.featureId == featureId }
   }
-
-
 }
